requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
imageio-ffmpeg>=0.4.9
numpy>=1.24.0
